Store a pointer to the end of each size path list for faster access.
In most normal cases the path list corresponding to a given file size
is short and this doesn't make much difference.

In edge cases where the file set being scanned contains a very large
number of files of the same size this can produce a substantial
reduction in scan time.

With this, the path list structure could use a bit of cleaning up, but
that's separate work.