Fix DIRENT_HAS_TYPE to work on filesystems that don't support it
Some Linux filesystems do not store d_type in a way that can be
retrieved "for free" by readdir(). XFS is one significant example.
This optimization causes files on XFS to sometimes get skipped as
"not a file" because XFS returns DT_UNKNOWN instead of DT_REG.
This change falls back to stat() when readdir() doesn't return
a valid d_type and dupd now works as expected on XFS.