(github #7) Support the -I|--hardlink-is-unique option during scan.
This will result in skipping additional hard links (beyond the first
entry seen) during the scan.

Implementation note: inodes are tracked via a bloom filter which means
there is a probability that files may be skipped incorrectly. At worst
this means a duplicate will not be reported, so it fails safely in that case..

Note: If --hardlink-is-unique is given during scan, it cannot be repeated
during interactive operations (file|ls|dups|uniques). This is enforced.