Reduce virtual memory usage by allocating additional path bocks as needed.
This also removes the hardcoded limit on total path * files so there
is no need for the --avg-size option anymore.

dupd used to allocate a very large path block at startup. Now it
allocates a small one first and chains additional ones if needed.

(In reality, this doesn't actually reduce RAM consumption because the
large buffer only became resident as needed and if the space is
needed, the same amount will be allocated with these changes.)