--hardlink-is-unique now forces HDD mode (implies --hdd).
--hardlink-is-unique requires tracking the inodes which is only done in HDD
mode. Previously, in SSD mode dupd used a bloom filter for this. Now, it
forces HDD mode to leverage the inode tracking done there.

The benefit is that the libbloom dependency can now be removed,
simplifying the build experience. The drawback is a small performance
hit in the case where --hardlink-is-unique is used on an SSD but the
HDD code runs well on SSD and --hardlink-is-unique is itself an edge
case so it is a reasonable compromise.