Add files to the path block only after two are seen with same size.
Before, scanning added all files to the path block. For files which
ended up having a unique size these entries don't get used later.

Now, cache the first file path in the sizetree itself. If and when a
second file of that size is found, only then add both to the path block.