Make HDD mode the default.
SSD mode can be selected with --ssd.

The HDD mode has improved to the point that it is a better default.
While it is sometimes slower than the SSD mode, it is never horribly
slow. The SSD mode on the other hand, while often faster, can have
extremely slow worst-case behavior. For that reason, the HDD mode makes for
a safer default.