Support for limiting read buffer allocation during HDD round1.
Default limit is 60% of available RAM. Can (but very rarely should)
be changed with --buflimit option.

The limit is rarely relevant but can be reached in cases where there
are many size sets with large numbers of files in each and the on-disk
layout is such that no size sets tend to get completed because the
reading is scattered amongst all of the sets. If the limit is
exceeded, reading switches to sequential mode to flush out size sets
(effectively, it goes into SSD mode).

The limit is not relevant in SSD mode because there it is already
reading sequentially and uses very little buffer memory.