(github #18): Statically regenerate embedded man page.
Instead of dynamically generating usage.o (containing the man page)
during the build, generate it via the 'man' target which is run
manually only when needed. The resulting file is checked in so it
doesn't need to be generated by the build.

Less elegant on one hand, but avoids the finicky opjcopy/ld conversion
from text file to object file which has proved troublesome to maintain
in a platform-independent way.