Removed SSD scan implementation and everything related.
Unfortunate because the SSD scan is both more memory efficient and
faster on SSD hardware.

However, the HDD mode has progressed enough that the SSD mode isn't
that much faster when it's faster. And it can be a lot slower when
it is slower.

The two implementations had diverged enough that continuing to
maintain both for all use cases was becoming more work than it was
worth given the fairly small performance benefit.