When a large file has been full hashed, save it in the cache.
Previously the hash was only saved if the files were duplicates, now
the hash is saved as soon as it is computed.

This is only makes a difference if the large files differ only in the
final block (dupd buffer block, not disk block) read from disk so it
is not a common case. But for such sets, we now get the hash for free
next time.